Agregar rol por cada menú y submenú * Se agrega campo rol_id en tabla menú y submenu para la relación con el rol * Se elimina archivo SubmenuSeeder.php y ViewsDefault.php * Se agregan los roles a los registros del menú para insertarlos a la bd en el archivo MenuSeeder.php * Se cambia lógica de cómo se ligan las vistas dependiendo del rol del usuario que se registra * Se agregan funciones en el repositorio de menu y submenu para buscar por roles

+++ database/migrations/2022_09_22_153502_create_submenu_user_table.php
@@ -14,7 +14,7 @@ class CreateSubmenuUserTable extends Migration
     public function up()
     {
         Schema::create('submenu_user', function (Blueprint $table) {
-            $table->unsignedInteger('submenu_id');
+            $table->unsignedBigInteger('submenu_id');
             $table->foreign('submenu_id')
                 ->references('id')
                 ->on('submenus');
